+++ src/main/java/com/github/terrapaw/terragrunt/inspection/TerragruntDeprecatedAttributeInspection.java
@@ -7,7 +7,7 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
 
-public class TerragruntDeprecatedAttributeInspection extends LocalInspectionTool {
+public class TerragruntDeprecatedAttributeInspection extends TerragruntBaseInspection {
     @NotNull
     @Override
     public PsiElementVisitor buildVisitor(@NotNull ProblemsHolder holder, boolean isOnTheFly) {