+++ src/main/java/com/github/terrapaw/terragrunt/inspection/TerragruntUnknownBlockInspection.java
@@ -8,7 +8,7 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import org.jetbrains.annotations.NotNull;
 
-public class TerragruntUnknownBlockInspection extends LocalInspectionTool {
+public class TerragruntUnknownBlockInspection extends TerragruntBaseInspection {
     @NotNull
     @Override
     public PsiElementVisitor buildVisitor(@NotNull ProblemsHolder holder, boolean isOnTheFly) {